Split up TestUTestMacro.cpp into two because of
64k segment limitation under Dos
